Update tree types
